minor ui tweaks. updated readme.md.
