+++ README.md
@@ -15,6 +15,7 @@ An open-source chat application to learn real-time messaging.
 ## .env.local file to be added
 
 ```
+NEXT_PUBLIC_MESSAGING=NEXT_PUBLIC_MESSAGING[sockets]
 NEXT_PUBLIC_LOCAL_STORAGE=NEXT_PUBLIC_LOCAL_STORAGE[local|cloud]
 NEXT_PUBLIC_SUPABASE_URL=YOUR_NEXT_PUBLIC_SUPABASE_URL
 NEXT_PUBLIC_SUPABASE_ANON_KEY=YOUR_NEXT_PUBLIC_SUPABASE_ANON_KEY
